% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{scheduler}
\alias{scheduler}
\title{The famous Jumping Rivers Scheduler}
\usage{
scheduler(rmd_name, repeats = 5, time_between = 5, params = NULL)
}
\arguments{
\item{rmd_name}{A Rmd file}

\item{repeats}{An integer. The number of times to call \code{render}}

\item{time_between}{The time (in seconds) between renders}

\item{params}{Parameters to be passed to the render script}
}
\description{
A simple schedular that runs \code{rmarkdown::render()}, the number 
of times equal to \code{reports}. The time interval between calls
is given by \code{time_between}
}
